src/backend/requirements.txt
Flask==3.0.0
google-cloud-datastore==2.15.1
requests~=2.31.0
python-jose
python-dotenv==1.0.1